+++ CHANGELOG.md
@@ -2,7 +2,17 @@
 
 All notable changes to this project will be documented in this file. Take a look at [the migration guide](docs/Migration%20Guide.md) to upgrade between two major versions.
 
-<!-- ## [Unreleased] -->
+## [Unreleased]
+
+### Added
+
+#### Navigator
+
+* Added `DragPointerObserver` to recognize drag gestures with pointer events.
+* Added `DirectionalNavigationAdapter.onNavigation` callback to be notified when a navigation action is triggered.
+    * This callback is called before executing any navigation action.
+    * Useful for hiding UI elements when the user navigates, or implementing analytics.
+
 
 ## [3.5.0]